refactor: extract scrollInputVisible into viewportState to eliminate duplication
The identical scrollInputVisible() + handleFocus() pair was copy-pasted
into InputModal, TextInputModal, and LoadBuildModal. Move it to
viewportState.ts as an exported utility so all three modals share a
single source of truth.
